working gift-wrapping

+++ include/Vector.hpp
@@ -6,16 +6,16 @@
 
 class Vector{
 private:
-    int x, y, z;
+    double x, y, z;
 public:
     Vector();
-    Vector(int x_coordinate, int y_coordinate, int z_coordinate);
-    int X() const;
-    int Y() const;
-    int Z() const;
+    Vector(double x_coordinate, double y_coordinate, double z_coordinate);
+    double X() const;
+    double Y() const;
+    double Z() const;
     bool operator==(const Vector &that) const;
     Vector operator*(const Vector &that) const;     //cross product
-    int dot(const Vector &that) const;        //dot product
+    double dot(const Vector &that) const;        //dot product
     double magnitude();
     Vector(const Vector &v) {x = v.x; y = v.y; z = v.z;} 
     Vector& operator = (const Vector &v) {
